[spec export] Change zoom-in range to 25%-400%
It's very useful to be able to zoom in closer in the spec export without having to resort to using the browser zoom function.

This change enables zooming from 25% up to 400% in spec export by changing the scale caps. Also tweaked the zoom-in increment when zooming above to make it more natural.

Disregard this proposed change if there is some deeper reason for capping the zoom where it was, but still consider finding a way to raise it to at least 200%. 100% seems very low and I'm in favor of letting the user zoom as much or as little as they want.